Refactoring SpeakerSocialLinks into a separate component
