+++ shared/components/Speaker/index.js
@@ -1,69 +1,21 @@
 import React, { PropTypes } from 'react';
+import SpeakerSocialLinks from '../SpeakerSocialLinks';
 
-const IconLink = ({ present, url, type, title }) => {
-  if (!present) { return null; }
-  return (
-    <a
-      className={'Speaker__detail Speaker__detail--' + type.toLowerCase()}
-      target="_blank"
-      rel="noopener"
-      href={url}
-      title={title}
-    />
-  );
-};
 
-IconLink.propTypes = {
-  present: PropTypes.string,
-  url: PropTypes.string,
-  type: PropTypes.string.isRequired,
-  title: PropTypes.string.isRequired,
-};
-
-const Speaker = ({
-  name,
-  company,
-  imageURL,
-  githubHandle,
-  twitterHandle,
-  blogURL,
-}) => {
-  const avatar = imageURL || '/img/PNG/Speaker_Anon.png';
+const Speaker = (props) => {
+  const avatar = props.imageURL || '/img/PNG/Speaker_Anon.png';
   return (
     <article className="Speaker">
       <figure className="Speaker__photo">
-        <img className="Speaker__photo--img" src={avatar} alt={name} />
+        <img className="Speaker__photo--img" src={avatar} alt={props.name} />
       </figure>
       <h5 className="Speaker__name">
-        <div className="Speaker__name--bold">{name} </div>
-        {company}
+        <div className="Speaker__name--bold">{props.name} </div>
+        {props.company}
       </h5>
-      <ul className="Speaker__details">
-        <li>
-          <IconLink
-            present={githubHandle}
-            type="GitHub"
-            url={'https://github.com/' + githubHandle}
-            title={name + "'s GitHub profile"}
-          />
-        </li>
-        <li>
-          <IconLink
-            present={twitterHandle}
-            type="Twitter"
-            url={'https://twitter.com/' + twitterHandle}
-            title={name + "'s Twitter profile"}
-          />
-        </li>
-        <li>
-          <IconLink
-            present={blogURL}
-            type="Blog"
-            url={blogURL}
-            title={name + "'s website"}
-          />
-        </li>
-      </ul>
+      <SpeakerSocialLinks
+        {...props}
+      />
 
     </article>
   );
